+++ include/http-parser.h
@@ -13,7 +13,7 @@ enum HTTP_PARSER_MODE {
 /// get/set maximum body size(global setting)
 /// @param[in] bytes 0-unlimited, other-limit bytes
 int http_get_max_size();
-int http_set_max_size(unsigned int bytes);
+int http_set_max_size(size_t bytes);
 
 /// create
 /// @param[in] mode 1-server mode, 0-client mode
@@ -31,7 +31,7 @@ void http_parser_clear(void* parser);
 /// @param[in] data content
 /// @param[in/out] bytes out-remain bytes
 /// @return 1-need more data, 0-receive done, <0-error
-int http_parser_input(void* parser, const void* data, int *bytes);
+int http_parser_input(void* parser, const void* data, size_t *bytes);
 
 /// HTTP start-line
 int http_get_version(void* parser, int *major, int *minor);
